requirements.txt
amqp==5.1.1
APScheduler==3.10.0
asgiref==3.6.0
async-timeout==4.0.2
atatus==1.4.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.1
Cython==0.29.33
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-celery-results==2.5.0
django-crispy-forms==2.0
django-grappelli==3.0.4
django-guardian==2.4.0
django-mmc==1.6.5
django-restframework==0.0.1
django-telegram-login==0.2.3
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
Faker==17.0.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
kombu==5.2.4
mysqlclient==2.1.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
pandas==1.5.3
Pillow==9.4.0
prompt-toolkit==3.0.38
psycopg2==2.9.1
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.0
sqlparse==0.4.3
treads==1.0.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
